refactor rxjava-common-3.0 packages to follow the changed module name
