+++ instrumentation/rxjava/rxjava-3.0/javaagent/src/test/java/RxJava3InstrumentationWithSpanTest.java
@@ -3,9 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import io.opentelemetry.instrumentation.rxjava.v3.common.AbstractRxJava3WithSpanTest;
-import io.opentelemetry.instrumentation.rxjava.v3.common.AbstractTracedWithSpan;
-import io.opentelemetry.instrumentation.rxjava.v3.common.extensionannotation.TracedWithSpan;
+import io.opentelemetry.instrumentation.rxjava.common.v3_0.AbstractRxJava3WithSpanTest;
+import io.opentelemetry.instrumentation.rxjava.common.v3_0.AbstractTracedWithSpan;
+import io.opentelemetry.instrumentation.rxjava.common.v3_0.extensionannotation.TracedWithSpan;
 import io.opentelemetry.instrumentation.testing.junit.AgentInstrumentationExtension;
 import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
 import org.junit.jupiter.api.extension.RegisterExtension;
