+++ instrumentation/rxjava/rxjava-3.1.1/library/src/main/java/io/opentelemetry/instrumentation/rxjava/v3_1_1/TracingAssembly.java
@@ -26,10 +26,10 @@ import io.opentelemetry.context.Context;
 import io.opentelemetry.context.Scope;
 import io.opentelemetry.instrumentation.api.annotation.support.async.AsyncOperationEndStrategies;
 import io.opentelemetry.instrumentation.api.internal.GuardedBy;
-import io.opentelemetry.instrumentation.rxjava.v3.common.RxJava3AsyncOperationEndStrategy;
-import io.opentelemetry.instrumentation.rxjava.v3.common.TracingCompletableObserver;
-import io.opentelemetry.instrumentation.rxjava.v3.common.TracingMaybeObserver;
-import io.opentelemetry.instrumentation.rxjava.v3.common.TracingSingleObserver;
+import io.opentelemetry.instrumentation.rxjava.common.v3_0.RxJava3AsyncOperationEndStrategy;
+import io.opentelemetry.instrumentation.rxjava.common.v3_0.TracingCompletableObserver;
+import io.opentelemetry.instrumentation.rxjava.common.v3_0.TracingMaybeObserver;
+import io.opentelemetry.instrumentation.rxjava.common.v3_0.TracingSingleObserver;
 import io.reactivex.rxjava3.core.Completable;
 import io.reactivex.rxjava3.core.CompletableObserver;
 import io.reactivex.rxjava3.core.Flowable;
