+++ instrumentation/rxjava/rxjava-common-3.0/library/src/main/java/io/opentelemetry/instrumentation/rxjava/common/v3_0/TracingMaybeObserver.java
@@ -20,21 +20,21 @@
  * under the License.
  */
 
-package io.opentelemetry.instrumentation.rxjava.v3.common;
+package io.opentelemetry.instrumentation.rxjava.common.v3_0;
 
 import io.opentelemetry.context.Context;
 import io.opentelemetry.context.Scope;
-import io.reactivex.rxjava3.core.SingleObserver;
+import io.reactivex.rxjava3.core.MaybeObserver;
 import io.reactivex.rxjava3.disposables.Disposable;
 import io.reactivex.rxjava3.internal.disposables.DisposableHelper;
 
-public class TracingSingleObserver<T> implements SingleObserver<T>, Disposable {
+public class TracingMaybeObserver<T> implements MaybeObserver<T>, Disposable {
 
-  private final SingleObserver<T> actual;
+  private final MaybeObserver<T> actual;
   private final Context context;
   private Disposable disposable;
 
-  public TracingSingleObserver(SingleObserver<T> actual, Context context) {
+  public TracingMaybeObserver(MaybeObserver<T> actual, Context context) {
     this.actual = actual;
     this.context = context;
   }
@@ -44,7 +44,7 @@ public class TracingSingleObserver<T> implements SingleObserver<T>, Disposable {
     if (!DisposableHelper.validate(disposable, d)) {
       return;
     }
-    this.disposable = d;
+    disposable = d;
     actual.onSubscribe(this);
   }
 
@@ -56,9 +56,16 @@ public class TracingSingleObserver<T> implements SingleObserver<T>, Disposable {
   }
 
   @Override
-  public void onError(Throwable throwable) {
+  public void onError(Throwable e) {
     try (Scope ignored = context.makeCurrent()) {
-      actual.onError(throwable);
+      actual.onError(e);
+    }
+  }
+
+  @Override
+  public void onComplete() {
+    try (Scope ignored = context.makeCurrent()) {
+      actual.onComplete();
     }
   }
