+++ instrumentation/rxjava/rxjava-common-3.0/library/src/test/java/RxJava3AsyncOperationEndStrategyTest.java
@@ -15,7 +15,7 @@ import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.instrumentation.api.annotation.support.async.AsyncOperationEndStrategy;
 import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
-import io.opentelemetry.instrumentation.rxjava.v3.common.RxJava3AsyncOperationEndStrategy;
+import io.opentelemetry.instrumentation.rxjava.common.v3_0.RxJava3AsyncOperationEndStrategy;
 import io.reactivex.rxjava3.core.Completable;
 import io.reactivex.rxjava3.core.Flowable;
 import io.reactivex.rxjava3.core.Maybe;
