+++ instrumentation/rxjava/rxjava-common-3.0/testing/src/main/java/io/opentelemetry/instrumentation/rxjava/common/v3_0/AbstractTracedWithSpan.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.rxjava.v3.common;
+package io.opentelemetry.instrumentation.rxjava.common.v3_0;
 
 import io.reactivex.rxjava3.core.Completable;
 import io.reactivex.rxjava3.core.Flowable;
