+++ instrumentation/rxjava/rxjava-common-3.0/testing/src/main/java/io/opentelemetry/instrumentation/rxjava/common/v3_0/RxJava3ConcurrencyTestHelper.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.rxjava.v3.common;
+package io.opentelemetry.instrumentation.rxjava.common.v3_0;
 
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
